optimisation getmobility, pruning a0.7, time a0.2

+++ Chess/src/org/elvio/chess/elements/HashBoard.java
@@ -7,7 +7,7 @@ import org.elvio.chess.util.BoardEvalue;
 
 public class HashBoard {
 
-	private HashMap<String, Double> boards;	
+	private HashMap<String, Integer> boards;	
 	private final static long MAX_BOARD = 1000000;
 	private long compteur;
 	
@@ -16,12 +16,12 @@ public class HashBoard {
 		compteur = 0;
 	}
 	
-	public Double getScore(Board board){
-		Double score = getScore(board.toLong());
+	public Integer getScore(Board board){
+		Integer score = getScore(board.toLong());
 		return score;
 	}
 	
-	private Double getScore(long[] board){
+	private Integer getScore(long[] board){
 		return boards.get(Arrays.toString(board));
 	}
