+++ Chess/src/org/elvio/chess/eval/algo/Materiel1.java
@@ -27,7 +27,7 @@ public class Materiel1  implements FonctionEvaluation {
 	private final static int cavalier_n = -300;
 	private final static int pion_n = -100;
 	
-	public final double getEval(Board board, int cpt){
+	public final int getEval(Board board, int cpt){
 		int resultat = 0;
 		for(int position = 0 ; position < BoardUtils.NBRE_CASES_BOARD ; position++){
 			resultat += eval(board.get2(position));
